+++ src/httptaskserver/EpicHandler.java
@@ -1,15 +1,15 @@
-package HttpTaskServer;
+package httptaskserver;
 
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 import manager.NotFoundException;
 import manager.TaskManager;
-import model.Subtask;
+import model.Epic;
 
 import java.io.IOException;
 
-class SubtaskHandler extends BaseHttpHandler implements HttpHandler {
-    public SubtaskHandler(TaskManager manager) {
+class EpicHandler extends BaseHttpHandler implements HttpHandler {
+    public EpicHandler(TaskManager manager) {
         super(manager);
     }
 
@@ -19,7 +19,7 @@ class SubtaskHandler extends BaseHttpHandler implements HttpHandler {
         String uri = exchange.getRequestURI().getPath();
         String[] uriParts = uri.split("/");
         if (method.equals("GET") && uriParts.length == 2) {
-            String text = gson.toJson(manager.getSubtasks());
+            String text = gson.toJson(manager.getEpics());
             sendText(exchange, text, method);
         } else if (method.equals("GET") && uriParts.length == 3) {
             try {
@@ -31,23 +31,19 @@ class SubtaskHandler extends BaseHttpHandler implements HttpHandler {
             }
         } else if (method.equals("POST") && uriParts.length == 2) {
             String input = new String(exchange.getRequestBody().readAllBytes(), DEFAULTCHARSET);
-            Subtask subtask = gson.fromJson(input, Subtask.class);
-            if (manager.isOverlap(subtask)) {
-                sendHasInteractions(exchange);
-                return;
-            }
-            if (subtask.getId() == 0) {
-                manager.addSubtask(subtask);
-                sendText(exchange, "Task has been successfully added", "POST");
+            Epic epic = gson.fromJson(input, Epic.class);
+            if (epic.getId() == 0) {
+                manager.addEpic(epic);
+                sendText(exchange, "Epic has been successfully added", "POST");
             } else {
-                manager.updateSubtask(subtask);
-                sendText(exchange, "Task has been successfully added", "POST");
+                manager.updateEpic(epic);
+                sendText(exchange, "Epic has been successfully added", "POST");
             }
         } else if (method.equals("DELETE") && uriParts.length == 3) {
             try {
                 int id = Integer.parseInt(uriParts[2]);
                 manager.removeById(id);
-                String text = "The task has been deleted";
+                String text = "The epic has been deleted";
                 sendText(exchange, text, method);
             } catch (NumberFormatException | NotFoundException e) {
                 sendNotFound(exchange);
@@ -57,3 +53,4 @@ class SubtaskHandler extends BaseHttpHandler implements HttpHandler {
         }
     }
 }
+
