+++ src/httptaskserver/LocalDateTimeAdapter.java
@@ -1,4 +1,4 @@
-package HttpTaskServer;
+package httptaskserver;
 
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
