+++ src/httptaskserver/TaskHandler.java
@@ -1,15 +1,15 @@
-package HttpTaskServer;
+package httptaskserver;
 
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 import manager.NotFoundException;
 import manager.TaskManager;
-import model.Epic;
+import model.Task;
 
 import java.io.IOException;
 
-class EpicHandler extends BaseHttpHandler implements HttpHandler {
-    public EpicHandler(TaskManager manager) {
+class TaskHandler extends BaseHttpHandler implements HttpHandler {
+    public TaskHandler(TaskManager manager) {
         super(manager);
     }
 
@@ -19,7 +19,7 @@ class EpicHandler extends BaseHttpHandler implements HttpHandler {
         String uri = exchange.getRequestURI().getPath();
         String[] uriParts = uri.split("/");
         if (method.equals("GET") && uriParts.length == 2) {
-            String text = gson.toJson(manager.getEpics());
+            String text = gson.toJson(manager.getTasks());
             sendText(exchange, text, method);
         } else if (method.equals("GET") && uriParts.length == 3) {
             try {
@@ -31,19 +31,23 @@ class EpicHandler extends BaseHttpHandler implements HttpHandler {
             }
         } else if (method.equals("POST") && uriParts.length == 2) {
             String input = new String(exchange.getRequestBody().readAllBytes(), DEFAULTCHARSET);
-            Epic epic = gson.fromJson(input, Epic.class);
-            if (epic.getId() == 0) {
-                manager.addEpic(epic);
-                sendText(exchange, "Epic has been successfully added", "POST");
+            Task task = gson.fromJson(input, Task.class);
+            if (manager.isOverlap(task)) {
+                sendHasInteractions(exchange);
+                return;
+            }
+            if (task.getId() == 0) {
+                manager.addTask(task);
+                sendText(exchange, "Task has been successfully added", "POST");
             } else {
-                manager.updateEpic(epic);
-                sendText(exchange, "Epic has been successfully added", "POST");
+                manager.updateTask(task);
+                sendText(exchange, "Task has been successfully added", "POST");
             }
         } else if (method.equals("DELETE") && uriParts.length == 3) {
             try {
                 int id = Integer.parseInt(uriParts[2]);
                 manager.removeById(id);
-                String text = "The epic has been deleted";
+                String text = "The task has been deleted";
                 sendText(exchange, text, method);
             } catch (NumberFormatException | NotFoundException e) {
                 sendNotFound(exchange);
@@ -53,4 +57,3 @@ class EpicHandler extends BaseHttpHandler implements HttpHandler {
         }
     }
 }
-
