+++ test/HttpTaskManagerTasksTest.java
@@ -1,4 +1,4 @@
-import HttpTaskServer.*;
+import httptaskserver.*;
 import com.google.gson.*;
 import manager.InMemoryTaskManager;
 import manager.TaskManager;
